compiler: Add heuristics to improve fusion lowering turnaround
