+++ devito/passes/clusters/__init__.py
@@ -3,6 +3,7 @@ from .buffering import *  # noqa
 from .cse import *  # noqa
 from .aliases import *  # noqa
 from .factorization import *  # noqa
+from .fusion import *  # noqa
 from .blocking import *  # noqa
 from .asynchrony import *  # noqa
 from .implicit import *  # noqa
